+++ R/graph.curve.params.R
@@ -3,19 +3,20 @@
 #' @param object
 #' a [data.frame()] containing MCMC samples of antibody decay curve parameters
 #' @param verbose verbose output
-#' @param show_all_curves whether to show individual curves under quantiles
 #' @param antigen_isos antigen isotypes to analyze
 #' (can subset `object`)
 #' @param alpha_samples `alpha` parameter passed to [ggplot2::geom_line]
-#' (has no effect if `show_all_curves = FALSE`)
+#' (has no effect if `iters_to_graph` is empty)
 #' @param show_quantiles whether to show point-wise (over time) quantiles
 #' @param log_x should the x-axis be on a logarithmic scale (`TRUE`)
 #' or linear scale (`FALSE`, default)?
 #' @param log_y should the Y-axis be on a logarithmic scale
 #' (default, `TRUE`) or linear scale (`FALSE`)?
+#' @inheritParams plot_curve_params_one_ab
 #' @param ... not currently used
 #' @returns a [ggplot2::ggplot()] object
 #' @export
+#' @inherit plot_curve_params_one_ab details
 #'
 #' @examples
 #' curve <-
@@ -26,7 +27,7 @@
 #'
 #' print(plot1)
 #'
-#' plot2 <- graph.curve.params(curve, show_all_curves = TRUE)
+#' plot2 <- graph.curve.params(curve, n_curves = 100)
 #' show(plot2)
 #'
 graph.curve.params <- function( # nolint: object_name_linter
@@ -34,10 +35,11 @@ graph.curve.params <- function( # nolint: object_name_linter
   antigen_isos = unique(object$antigen_iso),
   verbose = FALSE,
   show_quantiles = TRUE,
-  show_all_curves = FALSE,
   alpha_samples = 0.3,
   log_x = FALSE,
   log_y = TRUE,
+  n_curves = 100,
+  iters_to_graph = object$iter |> unique() |> head(n_curves),
   ...
 ) {
   if (verbose) {
@@ -51,28 +53,6 @@ graph.curve.params <- function( # nolint: object_name_linter
     dplyr::filter(.data$antigen_iso %in% antigen_isos)
 
   tx2 <- 10^seq(-1, 3.1, 0.025)
-
-  bt <- function(y0, y1, t1) {
-    log(y1 / y0) / t1
-  }
-
-  # uses r > 1 scale for shape
-  ab <- function(t, y0, y1, t1, alpha, shape) {
-    beta <- bt(y0, y1, t1)
-
-    yt <- 0
-
-    if (t <= t1) {
-      yt <- y0 * exp(beta * t)
-    }
-
-    if (t > t1) {
-      yt <- (y1^(1 - shape) - (1 - shape) * alpha * (t - t1))^(1 / (1 - shape))
-    }
-
-    return(yt)
-  }
-
 
   d <- object
 
@@ -100,7 +80,7 @@ graph.curve.params <- function( # nolint: object_name_linter
     ) |>
     select(-"name") |>
     rowwise() |>
-    mutate(res = ab(
+    mutate(res = ab1(
       .data$t,
       .data$y0,
       .data$y1,
@@ -142,8 +122,8 @@ graph.curve.params <- function( # nolint: object_name_linter
     serocourse_sum |>
     dplyr::filter(.data$quantile %in% c("med", "p10", "p90")) |>
     dplyr::summarize(
-      min = min(.data$res, 0.9),
-      max = max(.data$res, 2000)
+      min = min(.data$res),
+      max = max(.data$res)
     )
 
 
@@ -158,38 +138,48 @@ graph.curve.params <- function( # nolint: object_name_linter
     ) +
     ggplot2::theme_minimal() +
     ggplot2::theme(axis.line = ggplot2::element_line()) +
-    ggplot2::labs(x = "Days since fever onset",
-                  y = "ELISA units",
-                  col = if_else(show_all_curves, "MCMC chain", "")) +
+    ggplot2::labs(
+      x = "Days since fever onset",
+      y = "ELISA units",
+      col = if_else(
+        length(iters_to_graph) > 0,
+        "MCMC chain",
+        ""
+      )
+    ) +
     ggplot2::theme(legend.position = "bottom")
 
-  if (show_all_curves) {
+  if (length(iters_to_graph) > 0) {
+
+    sc_to_graph <-
+      serocourse_all |>
+      filter(.data$iter %in% iters_to_graph)
 
     range <-
-      serocourse_all |>
+      sc_to_graph |>
       dplyr::summarize(
-        min = min(.data$res, 0.9),
-        max = max(.data$res, 2000)
+        min = min(.data$res),
+        max = max(.data$res)
       )
 
     group_vars <-
       c("iter", "chain") |>
-      intersect(names(serocourse_all))
+      intersect(names(sc_to_graph))
 
     if (length(group_vars) > 1) {
-      serocourse_all <-
-        serocourse_all |>
+      sc_to_graph <-
+        sc_to_graph |>
         mutate(
-          iter = interaction(across(all_of(group_vars)))
+          group = interaction(across(all_of(group_vars)))
         )
       plot1 <-
         plot1 +
         geom_line(
-          data = serocourse_all,
+          data = sc_to_graph,
           alpha = alpha_samples,
           aes(
             color = .data$chain |> factor(),
-            group = .data$iter
+            group = .data$group
           )
         ) +
         ggplot2::expand_limits(y = range)
@@ -197,7 +187,7 @@ graph.curve.params <- function( # nolint: object_name_linter
 
       plot1 <-
         plot1 +
-        geom_line(data = serocourse_all,
+        geom_line(data = sc_to_graph,
                   alpha = alpha_samples,
                   aes(group = .data$iter)) +
         ggplot2::expand_limits(y = range)