+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## New features
 
+* Made `graph.curve.params()` the default sub-method for `autoplot.curve_params()` (#450)
 * Added `log_x` and `log_y` options to `graph.curve.params()` sub-method for 
 `autoplot.curve_params()` (#453)
 * Extended `sim_pop_data_multi()` to loop over multiple sample sizes (#444)